gcov: Avoid 16-bit sign extension error (coverity)

+++ src/include/gcov.hh
@@ -19,8 +19,8 @@ namespace kcov
 	static inline uint64_t gcovGetAddress(const std::string &filename, int32_t function,
 			int32_t basicBlock, int32_t lineIndex)
 	{
-		uint16_t fn16 = ((uint32_t)function) & 0xffff;
-		uint16_t bb16 = ((uint32_t)basicBlock) & 0xffff;
+		uint32_t fn16 = ((uint32_t)function) & 0xffff;
+		uint32_t bb16 = ((uint32_t)basicBlock) & 0xffff;
 		uint64_t fnAndBb = (fn16 << 16) | bb16;
 		uint64_t fileNameHash = (((uint64_t)std::hash<std::string>()(filename)) & 0xffffff00ULL);
 		uint64_t lineIndexMask = lineIndex & 0xffULL;
